SHL/requirements.txt
flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
huggingface_hub==0.16.4
pandas~=2.2.3
faiss-cpu~=1.10.0
numpy~=2.2.4
